refactor/move hash functions to common functions.ts file
